refactor(theme): split theme runtime and context into explicit module files
