+++ src/theme/ThemeProvider.tsx
@@ -1,59 +1,19 @@
-import {
-  createContext,
-  useCallback,
-  useContext,
-  useEffect,
-  useMemo,
-  useState,
-  type PropsWithChildren,
-} from 'react'
+import { useCallback, useEffect, useMemo, useState, type PropsWithChildren } from 'react'
 import {
   APP_THEME_OPTIONS,
-  APP_THEME_STORAGE_KEY,
-  DEFAULT_APP_THEME_ID,
   getAppTheme,
-  isAppThemeId,
-  type AppThemeDefinition,
   type AppThemeId,
 } from './themes'
-import { readStorageItem, writeStorageItem } from '@shared/lib'
-
-type ThemeContextValue = {
-  setTheme: (themeId: AppThemeId) => void
-  theme: AppThemeDefinition
-  themeId: AppThemeId
-  themes: AppThemeDefinition[]
-}
-
-const ThemeContext = createContext<ThemeContextValue | null>(null)
-
-function applyTheme(theme: AppThemeDefinition) {
-  if (typeof document === 'undefined') {
-    return
-  }
-
-  document.documentElement.dataset.theme = theme.id
-  document.documentElement.style.colorScheme = theme.colorScheme
-}
-
-function readInitialTheme() {
-  if (typeof window === 'undefined') {
-    return DEFAULT_APP_THEME_ID
-  }
-
-  const storedTheme = readStorageItem(APP_THEME_STORAGE_KEY)
-  return isAppThemeId(storedTheme) ? storedTheme : DEFAULT_APP_THEME_ID
-}
+import { applyAppTheme, persistAppThemeId, readSavedAppThemeId } from './runtime'
+import { ThemeContext, type ThemeContextValue } from './themeContext'
 
 export function ThemeProvider({ children }: PropsWithChildren) {
-  const [themeId, setThemeId] = useState<AppThemeId>(() => readInitialTheme())
+  const [themeId, setThemeId] = useState<AppThemeId>(() => readSavedAppThemeId())
   const theme = useMemo(() => getAppTheme(themeId), [themeId])
 
   useEffect(() => {
-    applyTheme(theme)
-    if (typeof window !== 'undefined') {
-      writeStorageItem(APP_THEME_STORAGE_KEY, themeId)
-    }
+    applyAppTheme(theme)
+    persistAppThemeId(themeId)
   }, [theme, themeId])
 
   const setTheme = useCallback((nextThemeId: AppThemeId) => {
@@ -73,14 +33,4 @@ export function ThemeProvider({ children }: PropsWithChildren) {
   )
 
   return <ThemeContext.Provider value={value}>{children}</ThemeContext.Provider>
-}
-
-export function useTheme() {
-  const context = useContext(ThemeContext)
-
-  if (!context) {
-    throw new Error('useTheme must be used within a ThemeProvider.')
-  }
-
-  return context
 }
