+++ src/theme/index.ts
@@ -1,11 +1,16 @@
-export { ThemeProvider, useTheme } from './ThemeProvider'
+export { ThemeProvider } from './ThemeProvider'
+export { useTheme } from './themeContext'
+export { APP_THEME_STORAGE_KEY } from './runtime'
 export {
   APP_THEME_DEFINITIONS,
   APP_THEME_OPTIONS,
-  APP_THEME_STORAGE_KEY,
   APP_THEMES,
   DEFAULT_APP_THEME_ID,
   getAppTheme,
   isAppThemeId,
 } from './themes'
-export type { AppThemeColorScheme, AppThemeDefinition, AppThemeId } from './themes'
+export type {
+  AppThemeColorScheme,
+  AppThemeDefinition,
+  AppThemeId,
+} from './themes'
